Add Memory usage measuring and MLE detection
Make use of the tool `timeout` that is mention in the Memory usage
Issue. Because of this, a custom Docker image has to be made. Other than
that, a test result now come with an extra element at the end that is
the amount of (max) memory it has used in kilobytes!